# api/openapi.yaml
openapi: 3.1.0
info:
  title: capp-backend API
  description: |
    Multi-cluster API gateway for managing Capp (container-app) resources on
    Kubernetes / OpenShift clusters.

    ## Authentication

    The backend supports five auth modes configured server-side:

    - **passthrough** — supply a raw Kubernetes `Bearer <token>` on every request.
    - **jwt** — call `/api/v1/auth/login` with `cluster` + `token` to exchange a
      cluster token for a short-lived JWT; refresh with `/api/v1/auth/refresh`.
    - **static** — use a fixed API key configured on the server (dev / CI only).
    - **dex** — call `/api/v1/auth/login` with `username` + `password`; Dex handles OIDC.
    - **openshift** — use `/api/v1/auth/openshift/authorize` to get an OAuth URL,
      then POST the code to `/api/v1/auth/openshift/callback`.

    Use `GET /api/v1/auth/mode` to discover which mode the server is running.
    All resource endpoints require the `Authorization: Bearer <token>` header.
  version: "0.1.0"
  contact:
    name: dana-team
    url: https://github.com/dana-team/capp-backend
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        In **passthrough** mode supply the raw Kubernetes ServiceAccount token.
        In **jwt** / **dex** / **openshift** modes supply the JWT returned by the
        login or callback endpoint.
        In **static** mode supply one of the configured API keys.
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: resource "my-capp" not found
            status:
              type: integer
              example: 404

    # ── Auth ─────────────────────────────────────────────────────────────────────
    LoginRequest:
      description: |
        - **jwt / passthrough / static** mode: provide `cluster` and `token`.
        - **dex** mode: provide `username` and `password`.
        - **openshift** mode: provide `username` and `password` (Resource Owner Password flow).
      oneOf:
        - title: ClusterTokenLogin
          type: object
          required: [cluster, token]
          properties:
            cluster:
              type: string
              description: Cluster to authenticate against (jwt/passthrough/static mode).
              example: production
            token:
              type: string
              description: Raw Kubernetes Bearer token (jwt/passthrough/static mode).
              example: "eyJhbGci..."
        - title: DexLogin
          type: object
          required: [username, password]
          properties:
            username:
              type: string
              description: Username (dex mode).
              example: admin
            password:
              type: string
              description: Password (dex mode).
              example: "s3cr3t"
        - title: OpenShiftPasswordLogin
          type: object
          required: [username, password]
          properties:
            username:
              type: string
              description: Username (openshift mode, Resource Owner Password flow).
              example: admin
            password:
              type: string
              description: Password (openshift mode).
              example: "s3cr3t"
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
          description: Short-lived JWT for API requests.
        refreshToken:
          type: string
          description: Long-lived JWT for refreshing the access token.
        expiresAt:
          type: string
          format: date-time
          description: UTC expiry time of the access token.
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          description: Refresh token obtained from a previous login or refresh.
...
    OpenShiftAuthorizeResponse:
      type: object
      properties:
        authorizeUrl:
          type: string
          description: OAuth authorize URL to redirect the browser to.
          example: "https://oauth.cluster.example.com/oauth/authorize?..."
        state:
          type: string
          description: CSRF state token to pass back in the callback request.
          example: "a1b2c3d4e5f6..."
    OpenShiftCallbackRequest:
      type: object
      required: [code, state]
      properties:
        code:
          type: string
          description: OAuth authorization code received from the OpenShift OAuth server.
        state:
          type: string
          description: CSRF state token received from the authorize response.

    # ── Clusters ──────────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Log in and obtain a token pair
      description: |
        - **jwt / passthrough / static** mode: provide `cluster` and `token`.
        - **dex** mode: provide `username` and `password`.

        Returns a short-lived access JWT and a long-lived refresh JWT.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        "400":
          description: Missing or malformed request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          description: Cluster not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Refresh an expired access token
      description: |
        Validates the supplied refresh token, extends the server-side session,
        and issues a new JWT pair. Not available in passthrough or static mode.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        "200":
          description: Token refreshed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/openshift/authorize:
    get:
      tags: [auth]
      summary: Get OpenShift OAuth authorize URL
      description: |
        Returns the OpenShift OAuth server's authorize URL. The frontend should
        redirect the browser to this URL to begin the OAuth flow.
        Only available when `auth.mode` is `openshift`.
      security: []
      responses:
        "200":
          description: Authorize URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenShiftAuthorizeResponse'
        "500":
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/openshift/callback:
    post:
      tags: [auth]
      summary: Exchange OpenShift OAuth code for tokens
      description: |
        Exchanges the OAuth authorization code (received after the browser redirect)
        for an access/refresh token pair. Only available when `auth.mode` is `openshift`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenShiftCallbackRequest'
      responses:
        "200":
          description: OAuth exchange successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        "400":
          description: Missing or malformed request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalError'

  # ── Clusters ───────────────────────────────────────────────────────────────────
...